fix(android): remove runtime deprecation warnings

+++ android/app/src/main/java/app/secpal/ProvisioningBootstrapRuntimeInfo.java
@@ -15,7 +15,7 @@ import androidx.core.content.pm.PackageInfoCompat;
 final class ProvisioningBootstrapRuntimeInfo {
     private final String packageName;
     private final String packageVersionName;
-    private final int packageVersionCode;
+    private final long packageVersionCode;
     private final String deviceName;
     private final String deviceManufacturer;
     private final String deviceModel;
@@ -24,7 +24,7 @@ final class ProvisioningBootstrapRuntimeInfo {
     ProvisioningBootstrapRuntimeInfo(
         String packageName,
         String packageVersionName,
-        int packageVersionCode,
+        long packageVersionCode,
         String deviceName,
         String deviceManufacturer,
         String deviceModel,
@@ -42,7 +42,7 @@ final class ProvisioningBootstrapRuntimeInfo {
     static ProvisioningBootstrapRuntimeInfo fromContext(Context context) {
         String packageName = context.getPackageName();
         String versionName = null;
-        int versionCode = 0;
+        long versionCode = 0;
 
         try {
             PackageManager packageManager = context.getPackageManager();
@@ -58,7 +58,7 @@ final class ProvisioningBootstrapRuntimeInfo {
             }
 
             versionName = packageInfo.versionName;
-            versionCode = (int) PackageInfoCompat.getLongVersionCode(packageInfo);
+            versionCode = PackageInfoCompat.getLongVersionCode(packageInfo);
         } catch (PackageManager.NameNotFoundException ignored) {
             // Fall back to package name only when package metadata is unavailable.
         }
@@ -82,7 +82,7 @@ final class ProvisioningBootstrapRuntimeInfo {
         return packageVersionName;
     }
 
-    int getPackageVersionCode() {
+    long getPackageVersionCode() {
         return packageVersionCode;
     }
